fix: Insert fails for accords with no corresponding case
Accords need the for_case column to be set to the caseId of the case
they belong to. So we mut check for a record with the caseId and
caseType and if it exists we set the for_case column to the id of the
existing record.

Insert fails otherwise.